requirements.txt
annotated-types==0.7.0
argcomplete==1.10.3
attrs==25.3.0
beautifulsoup4==4.8.2
blinker==1.9.0
blis==1.3.0
bs4==0.0.2
cachelib==0.13.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
compressed-rtf==1.0.7
confection==0.1.5
cryptography==45.0.4
cymem==2.0.11
docx2txt==0.9
ebcdic==1.1.1
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
extract-msg==0.28.7
filelock==3.13.1
Flask==3.1.1
Flask-Session==0.8.0
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.4
huggingface-hub==0.33.0
idna==3.10
IMAPClient==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
murmurhash==1.0.13
networkx==3.3
nltk==3.9.1
numpy==1.24.4
olefile==0.47
outcome==1.3.0.post0
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.0.0
preshed==3.0.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.26.1
PyPDF2==3.0.1
pypdfium2==4.30.1
PySocks==1.7.1
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.1.0
python-pptx==0.6.23
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
selenium==4.33.0
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.12.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
srsly==2.5.1
sympy==1.13.3
textract==1.6.5
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xlrd==1.2.0
xlsxwriter==3.2.4
